+++ generator/stdlib/src/commonTest/kotlin/com/github/jcornaz/kwik/generator/stdlib/BooleanGeneratorTest.kt
@@ -1,6 +1,7 @@
 package com.github.jcornaz.kwik.generator.stdlib
 
 import com.github.jcornaz.kwik.generator.api.Generator
+import com.github.jcornaz.kwik.generator.api.randomSequence
 import com.github.jcornaz.kwik.generator.test.AbstractGeneratorTest
 import kotlin.test.Test
 import kotlin.test.assertTrue
@@ -10,7 +11,7 @@ class BooleanGeneratorTest : AbstractGeneratorTest() {
 
     @Test
     fun produceTrueAndFalse() {
-        val values: Sequence<Boolean> = Generator.booleans().randoms(42).take(200)
+        val values: Sequence<Boolean> = Generator.booleans().randomSequence(42).take(200)
         assertTrue(values.any { it })
         assertTrue(values.any { !it })
     }