+++ generator/stdlib/src/commonTest/kotlin/com/github/jcornaz/kwik/generator/stdlib/DefaultGeneratorTest.kt
@@ -1,6 +1,7 @@
 package com.github.jcornaz.kwik.generator.stdlib
 
 import com.github.jcornaz.kwik.generator.api.Generator
+import com.github.jcornaz.kwik.generator.api.randomSequence
 import kotlin.test.Test
 import kotlin.test.assertFailsWith
 import kotlin.test.assertTrue
@@ -17,33 +18,33 @@ class DefaultGeneratorTest {
 
     @Test
     fun generateInts() {
-        assertTrue(Generator.default<Int>().randoms(0).first() is Int)
+        assertTrue(Generator.default<Int>().randomSequence(0).first() is Int)
     }
 
     @Test
     fun generateLongs() {
-        assertTrue(Generator.default<Long>().randoms(0).first() is Long)
+        assertTrue(Generator.default<Long>().randomSequence(0).first() is Long)
     }
 
     @Test
     fun generateFloats() {
-        assertTrue(Generator.default<Float>().randoms(0).first() is Float)
+        assertTrue(Generator.default<Float>().randomSequence(0).first() is Float)
     }
 
     @Test
     fun generateDoubles() {
-        assertTrue(Generator.default<Double>().randoms(0).first() is Double)
+        assertTrue(Generator.default<Double>().randomSequence(0).first() is Double)
     }
 
     @Test
     fun generateBooleans() {
-        val value = Generator.default<Boolean>().randoms(0).first()
+        val value = Generator.default<Boolean>().randomSequence(0).first()
         println(value)
         assertTrue(value is Boolean)
     }
 
     @Test
     fun generateStrings() {
-        assertTrue(Generator.default<String>().randoms(0).first() is String)
+        assertTrue(Generator.default<String>().randomSequence(0).first() is String)
     }
 }