+++ generator/test/src/commonMain/kotlin/com/github/jcornaz/kwik/generator/test/AbstractGeneratorTest.kt
@@ -1,6 +1,7 @@
 package com.github.jcornaz.kwik.generator.test
 
 import com.github.jcornaz.kwik.generator.api.Generator
+import com.github.jcornaz.kwik.generator.api.randomSequence
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNotEquals
@@ -10,22 +11,22 @@ abstract class AbstractGeneratorTest {
 
     @Test
     fun isPredictable() {
-        val generation1 = generator.randoms(1).take(200).toList()
-        val generation2 = generator.randoms(1).take(200).toList()
+        val generation1 = generator.randomSequence(1).take(200).toList()
+        val generation2 = generator.randomSequence(1).take(200).toList()
 
         assertEquals(generation1, generation2)
     }
 
     @Test
     fun isRandom() {
-        val generation1 = generator.randoms(0).take(200).toList()
-        val generation2 = generator.randoms(1).take(200).toList()
+        val generation1 = generator.randomSequence(0).take(200).toList()
+        val generation2 = generator.randomSequence(1).take(200).toList()
 
         assertNotEquals(generation1, generation2)
     }
 
     @Test
     fun isInfinite() {
-        assertEquals(100_000, generator.randoms(12).take(100_000).count())
+        assertEquals(100_000, generator.randomSequence(12).take(100_000).count())
     }
 }